backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
flask-mqtt==1.1.1
PyJWT==2.8.0
requests==2.31.0